Move things to notifier package and avoid redundant interface
