+++ pkg/services/ngalert/api/api.go
@@ -19,7 +19,6 @@ import (
 	"github.com/grafana/grafana/pkg/services/ngalert/notifier"
 	"github.com/grafana/grafana/pkg/services/ngalert/provisioning"
 	"github.com/grafana/grafana/pkg/services/ngalert/schedule"
-	"github.com/grafana/grafana/pkg/services/ngalert/services"
 	"github.com/grafana/grafana/pkg/services/ngalert/state"
 	"github.com/grafana/grafana/pkg/services/ngalert/store"
 	"github.com/grafana/grafana/pkg/services/quota"
@@ -92,7 +91,7 @@ func (api *API) RegisterAPIEndpoints(m *metrics.API) {
 	api.RegisterAlertmanagerApiEndpoints(NewForkedAM(
 		api.DatasourceCache,
 		NewLotexAM(proxy, logger),
-		&AlertmanagerSrv{secrets: api.SecretsService, log: logger, ac: api.AccessControl, configs: services.NewAlertmanagerConfigService(api.MultiOrgAlertmanager, api.SecretsService, api.AlertingStore, logger)},
+		&AlertmanagerSrv{secrets: api.SecretsService, log: logger, ac: api.AccessControl, configs: notifier.NewAlertmanagerConfigService(api.MultiOrgAlertmanager, api.SecretsService, api.AlertingStore, logger)},
 	), m)
 	// Register endpoints for proxying to Prometheus-compatible backends.
 	api.RegisterPrometheusApiEndpoints(NewForkedProm(
